Move mock message consumer from mock package to messagetest.
